user logs table to a new page
